Rename biome::users to biome::user
In most contexts, singular module names should be used.
