+++ libsplinter/src/biome/credentials/database/models.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use super::schema::*;
-use crate::biome::users::database::models::UserModel;
+use crate::biome::user::database::models::UserModel;
 
 #[derive(Queryable, Identifiable, Associations, PartialEq, Debug)]
 #[table_name = "user_credentials"]
